requirements.txt
pandas
gspread
oauth2client
openai
requests
python-dotenv
streamlit
